refactor: move tracking middleware
